fix: `protected( crate )` -> `protected` | move test to separate file
