+++ module/move/willbe/tests/inc/mod.rs
@@ -1,3 +1,4 @@
 use super::*;
 
 mod local_dependencies;
+mod publish_need;
